requirements.txt
## Requirements
python-dotenv
google-genai
alpaca-py
colorama


